# server/docs/swagger.yaml
basePath: /v1
definitions:
...
  github_com_buysell-technologies_summer-internship-2024-backend_api_handler_request.CreateUserRequest:
    properties:
      email:
        example: taro_tanaka@example.com
        type: string
      employee_number:
        example: "0000000000"
        maxLength: 10
        minLength: 1
        type: string
      gender:
        enum:
        - male
        - female
        example: male
        type: string
      name:
        example: 田中 太郎
        maxLength: 255
        minLength: 1
        type: string
      store_id:
        example: 00000000-0000-0000-0000-000000000000
        type: string
    required:
    - email
    - employee_number
    - name
    - store_id
    type: object
...
  github_com_buysell-technologies_summer-internship-2024-backend_api_handler_request.UpdateUserRequest:
    properties:
      email:
        example: taro_tanaka@example.com
        type: string
      employee_number:
        example: "0000000000"
        maxLength: 10
        minLength: 1
        type: string
      gender:
        enum:
        - male
        - female
        example: male
        type: string
      name:
        example: 田中 太郎
        maxLength: 255
        minLength: 1
        type: string
      store_id:
        example: 00000000-0000-0000-0000-000000000000
        type: string
    required:
    - email
    - employee_number
    - name
    - store_id
    type: object
...
  model.Order:
    properties:
      created_at:
        type: string
      customer_id:
        type: string
      delivery_date:
        type: string
      id:
        type: integer
      quantity:
        type: integer
      status:
        $ref: '#/definitions/model.OrderStatus'
      stock_id:
        type: integer
      total_amount:
        type: integer
      updated_at:
        type: string
    type: object
  model.OrderStatus:
    enum:
    - PENDING
    - SHIPPED
    - DELIVERED
    - CANCELLED
    type: string
    x-enum-comments:
      StatusCancelled: キャンセル
      StatusDelivered: 納品済み
      StatusPending: 保留中
      StatusShipped: 出荷済み
    x-enum-varnames:
    - StatusPending
    - StatusShipped
    - StatusDelivered
    - StatusCancelled
  model.Stock:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      orders:
        description: リレーション (hasMany)
        items:
          $ref: '#/definitions/model.Order'
        type: array
      price:
        type: integer
      quantity:
        type: integer
      store_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  model.User:
    properties:
      created_at:
        type: string
      deleted_at:
        example: "2023-01-01T00:00:00Z"
        format: date-time
        type: string
      email:
        type: string
      employee_number:
        type: string
      gender:
        type: string
      id:
        type: string
      name:
        type: string
      stocks:
        description: リレーション (hasMany)
        items:
          $ref: '#/definitions/model.Stock'
        type: array
      store_id:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:1234
info:
  contact: {}
  description: This is the API for the Summer Internship 2024 Backend.
  title: Summer Internship 2024 Backend API
  version: "1"
paths:
...
  /users:
    get:
      description: 従業員一覧の取得
      parameters:
      - description: 取得件数
        example: 10
        in: query
        minimum: 0
        name: limit
        type: integer
      - description: 取得開始位置
        example: 0
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: 従業員一覧の取得
    post:
      description: 従業員の作成
      parameters:
      - description: 作成条件
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/github_com_buysell-technologies_summer-internship-2024-backend_api_handler_request.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: 従業員の作成
  /users/{id}:
    delete:
      description: 従業員の削除
      parameters:
      - description: 従業員ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: 従業員の削除
    get:
      description: 従業員の取得
      parameters:
      - description: 従業員ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: 従業員の取得
    put:
      consumes:
      - application/json
      description: 従業員の更新
      parameters:
      - description: 従業員ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: 更新条件
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/github_com_buysell-technologies_summer-internship-2024-backend_api_handler_request.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: 従業員の更新
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"